refactor: AlarmApi 구조 리팩토링
