+++ src/main/kotlin/com/official/memento/alarm/domain/port/AlarmOutputPort.kt
@@ -1,0 +1,5 @@
+package com.official.memento.alarm.domain.port
+
+interface AlarmOutputPort {
+    fun sendAlarm(uri: String, content: String)
+}
